Remove usage legacy HTTP Client
Improved the "LookupUtils" to not use the deprecated HTTP classes
but "HttpURLConnection" instead. Contrary to the Apache HTTP client
the HttURLConnection is still existing in Android 6 and seems to be
more efficient besides being compatible with SSL SNI.

Furthermore set the missing "formattedNumber" attribute when the
"ContactBuilder" is creating the "ContactInfo" object after lookups.

Change-Id: I01be4fa09ceba7338aac63e2c718704fa1941ee5

Conflicts:
	Android.mk
	AndroidManifest.xml